cleaned up some frontend
